Small refactorings
