+++ example/src/main/java/com/b3nedikt/restring/example/BaseActivity.kt
@@ -5,11 +5,7 @@ import androidx.appcompat.app.AppCompatActivity
 
 import com.b3nedikt.restring.Restring
 
-/**
- * We should wrap the base context of our activities, which is better to put it
- * on BaseActivity, so that we don't have to repeat it for all activities one-by-one.
- */
-open class BaseActivity : AppCompatActivity() {
+abstract class BaseActivity : AppCompatActivity() {
 
     override fun attachBaseContext(newBase: Context) {
         super.attachBaseContext(Restring.wrapContext(newBase))
