+++ restring/src/main/java/com/b3nedikt/restring/RestringContextWrapper.kt
@@ -2,16 +2,23 @@ package com.b3nedikt.restring
 
 import android.content.Context
 import android.content.ContextWrapper
+import android.content.res.Resources
 import android.view.LayoutInflater
 
 /**
- * Main Restring context wrapper which wraps the context for providing another layout inflater & resources.
+ * Main Restring context wrapper which wraps the context for providing another layout inflater
+ * & resources.
  */
 internal class RestringContextWrapper private constructor(
         base: Context,
         stringRepository: StringRepository,
         private val viewTransformerManager: ViewTransformerManager
-) : ContextWrapper(CustomResourcesContextWrapper(base, RestringResources(base.resources, stringRepository))) {
+) : ContextWrapper(base) {
+
+    private val res: Resources by lazy {
+        val baseResources = super.getResources()
+        RestringResources(baseResources, stringRepository)
+    }
 
     private val layoutInflater: RestringLayoutInflater by lazy {
         RestringLayoutInflater(
@@ -29,6 +36,8 @@ internal class RestringContextWrapper private constructor(
 
         return super.getSystemService(name)
     }
+
+    override fun getResources() = res
 
     companion object {
